Unify macro source display
